requirements.txt
streamlit
pandas
nltk
joblib
psycopg2-binary
altair
plotly
matplotlib
geopandas
sklearn
tensorflow
numpy